# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BoilerPlate Express API with Swagger",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "Local server"
    },
    {
      "url": "https://staging.api-expressjs.boilerplate.hng.tech/",
      "description": "Live server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/verify-otp": {
      "post": {
        "summary": "Verify the user's email using OTP",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": {
                    "type": "integer",
                    "description": "The OTP sent to the user's email"
                  },
                  "token": {
                    "type": "string",
                    "description": "The token received during sign up"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid OTP or verification token has expired"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Some server error"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Log in a user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Some server error"
          }
        }
      }
    },
    "/api/v1/auth/forgotPassword": {
      "post": {
        "summary": "Request a password reset",
        "description": "Allows a user to request a password reset link by providing their email address.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "The email address of the user.",
                    "example": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully requested password reset.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "The status of the request.",
                      "example": "success"
                    },
                    "status_code": {
                      "type": "integer",
                      "description": "The HTTP status code.",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "description": "The message indicating the result of the request.",
                      "example": "Password reset link sent to your email."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/v1/auth/resetPassword": {
      "post": {
        "summary": "Reset a user's password",
        "description": "Allows a user to reset their password by providing a valid reset token and a new password.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "newPassword"],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "The password reset token.",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "The new password for the user.",
                    "example": "new_secure_password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully reset password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "The message indicating the result of the password reset.",
                      "example": "Password has been reset successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "patch": {
        "summary": "Change user password",
        "tags": ["Auth"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldPassword": {
                    "type": "string",
                    "description": "Current password of the user"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "New password to set for the user"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "description": "Confirmation of the new password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password changed successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, such as mismatched passwords or invalid input"
          },
          "401": {
            "description": "Unauthorized, invalid credentials or not authenticated"
          },
          "500": {
            "description": "Some server error"
          }
        }
      }
    },
    "/api/v1/auth/google-signin": {
      "post": {
        "summary": "Handle Google authentication and register/login a user",
        "description": "This endpoint handles Google OAuth2.0 authentication. It accepts a Google user payload and either registers a new user or logs in an existing one.",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The user's email address.",
                    "example": "user@example.com"
                  },
                  "email_verified": {
                    "type": "boolean",
                    "description": "Whether the user's email is verified.",
                    "example": true
                  },
                  "name": {
                    "type": "string",
                    "description": "The user's full name.",
                    "example": "John Doe"
                  },
                  "picture": {
                    "type": "string",
                    "format": "url",
                    "description": "URL to the user's profile picture.",
                    "example": "https://example.com/avatar.jpg"
                  },
                  "sub": {
                    "type": "string",
                    "description": "Google user ID (subject claim).",
                    "example": "1234567890"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User authenticated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Verify if authentication is successful",
                      "example": "Authentication successful"
                    },
                    "user": {
                      "type": "object",
                      "description": "The authenticated user object."
                    },
                    "access_token": {
                      "type": "string",
                      "description": "JWT access token for authentication.",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid or missing data in request body"
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred"
          }
        }
      }
    },
...
  },
...
}